Add options for user downloaded third-party libraries. User does not need re-complie the libraries if they has copy that to build/luojianet_ms/_deps
Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
	modified:   build.bat
	modified:   build.sh
	modified:   build_gdal.bat
	modified:   scripts/build/default_options.sh
	modified:   scripts/build/process_options.sh
	modified:   scripts/build/usage.sh
	modified:   version.txt

+++ build.bat
@@ -21,6 +21,7 @@ SET BUILD_PATH=%BASE_PATH%/build
 
 SET threads=8
 SET ENABLE_GITEE=ON
+SET ENABLE_THIRD=OFF
 
 set VERSION_STR=''
 for /f "tokens=1" %%a in (version.txt) do (set VERSION_STR=%%a)
@@ -43,33 +44,36 @@ IF NOT EXIST "%BUILD_PATH%/luojianet_ms" (
     md "luojianet_ms"
 )
 
-set PATCHES_FOLDER="%BUILD_PATH%/luojianet_ms/_ms_patch"
-echo %PATCHES_FOLDER%
 
-IF EXIST "%PATCHES_FOLDER%" (
-    rd %PATCHES_FOLDER% /S /Q
-)
+IF "%ENABLE_THIRD%" == "ON" (
+	set PATCHES_FOLDER="%BUILD_PATH%/luojianet_ms/_ms_patch"
+	echo %PATCHES_FOLDER%
 
-set DEPS_FOLDER="%BUILD_PATH%/luojianet_ms/_deps"
-IF EXIST "%DEPS_FOLDER%" (
-    FOR /D %%p IN ("%DEPS_FOLDER%\*-src") DO rmdir "%%p" /S /Q
-	FOR /D %%p IN ("%DEPS_FOLDER%\*-build") DO rmdir "%%p" /S /Q
-	FOR /D %%p IN ("%DEPS_FOLDER%\*-subbuild") do (
-		rmdir "%%p\CMakeFiles" /S /Q
-		del "%%p\CMake*" /S /Q
-		del "%%p\Make*" /S /Q
-		del "%%p\cmake*" /S /Q
-		echo "subpath: %%p\*-populate-prefix\src"
-		FOR /D %%m IN ("%%p\*-populate-prefix") DO (
-			FOR /D %%n IN ("%%m\src\*-stamp") DO rmdir "%%n" /S /Q
+	IF EXIST "%PATCHES_FOLDER%" (
+		rd %PATCHES_FOLDER% /S /Q
+	)
+
+	set DEPS_FOLDER="%BUILD_PATH%/luojianet_ms/_deps"
+	IF EXIST "%DEPS_FOLDER%" (
+		FOR /D %%p IN ("%DEPS_FOLDER%\*-src") DO rmdir "%%p" /S /Q
+		FOR /D %%p IN ("%DEPS_FOLDER%\*-build") DO rmdir "%%p" /S /Q
+		FOR /D %%p IN ("%DEPS_FOLDER%\*-subbuild") do (
+			rmdir "%%p\CMakeFiles" /S /Q
+			del "%%p\CMake*" /S /Q
+			del "%%p\Make*" /S /Q
+			del "%%p\cmake*" /S /Q
+			echo "subpath: %%p\*-populate-prefix\src"
+			FOR /D %%m IN ("%%p\*-populate-prefix") DO (
+				FOR /D %%n IN ("%%m\src\*-stamp") DO rmdir "%%n" /S /Q
+			)
 		)
+		FOR /D %%q IN ("%BUILD_PATH%\luojianet_ms") do (
+			echo "newpath: %%q"
+			del %%q\CMake* /S /Q
+			del %%q\cmake* /S /Q
+		)
+		REM FOR /D %%p IN ("%BUILD_PATH%\luojianet_ms\CMake*") DO rmdir "%%p" /S /Q
 	)
-	FOR /D %%q IN ("%BUILD_PATH%\luojianet_ms") do (
-		echo "newpath: %%q"
-		del %%q\CMake* /S /Q
-		del %%q\cmake* /S /Q
-	)
-	REM FOR /D %%p IN ("%BUILD_PATH%\luojianet_ms\CMake*") DO rmdir "%%p" /S /Q
 )
 
 cd %BUILD_PATH%/luojianet_ms
